# Optimize `CowData` and `LocalVector` functions `.insert` and `.remove_at` by using move semantics.