@hpvb 2024-12-19
I did some benchmarks using https://github.com/godotengine/godot-benchmarks between master and this pr + #100483 and after running both master and that combination several times I can confirm that emitting signals is roughly ~10% faster with these changes compared to master, I also think that I see an improvement in the ObjectDB benchmarks but with 64 threads on my computer there's a ton of noise in the threads and it is hard to be sure.

Function Callable seems to be 1% slower though, even after repeated runs. I can't explain it.