@mihe 2025-03-26
@Ivorforce Any particular reason why `LocalVector::push_back` did not get `std::move` added to it for non-trivially constructible objects?

https://github.com/godotengine/godot/blob/594d64ec244b6b99321e2096987d4d69de4c8845/core/templates/local_vector.h#L66-L70

Seeing as how this would apply to any object that defines pretty much any constructor, like `Variant` (which currently always seems to copy here), I would imagine there are some gains to be had there as well.

(It also happens to be the only thing standing in the way for `LocalVector` to support move-only types, as far as I can tell.)