@Ivorforce 2025-03-26
> @Ivorforce Any particular reason why `LocalVector::push_back` did not get `std::move` added to it for non-trivially constructible objects?

Nope! Just a mistake on my part. It should probably be added :)