@mihe 2025-03-27
> Nope! Just a mistake on my part. It should probably be added :)

I threw together a PR for it: #104693